pyproject.toml
# The preference build system would have been flit, but setuptools had to be
# used to support `pip install git+http...` within Databricks. When trying
# to parse the version using flit, it needs to import the full package
# and some dependencies might be missing in the virtual environment.
# Setuptools allow to explicitly set the location of the version without
# having to import the package.
[build-system]
requires = [
#    "flit_core >=3.2,<4",
    "setuptools",
    "setuptools-scm",
]
#build-backend = "flit_core.buildapi"
build-backend = "setuptools.build_meta"

[project]
name = "settus"
authors = [
    {name = "Olivier Soucy", email = "olivier.soucy@okube.ai"},
]
description = "Settings management using Pydantic, with ability to retrieve cloud secrets."
readme = "README.md"
requires-python = ">=3.9"
keywords = ["one", "two"]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Pydantic :: 2",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dynamic = ["version"]
dependencies = [
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "black",
#    "flit",
    "mkdocs",
    "mkdocstrings[python]",
    "mkdocs-material",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-examples",
#    "tox"
]
azure = [
    "azure-identity",
    "azure-keyvault"
]
aws = [
    "boto3",
]
gcp = [
]
databricks = [
]

[project.urls]
"Homepage" = "https://github.com/opencubes-ai/settus"
"Bug Tracker" = "https://github.com/opencubes-ai/settus/issues"

[tool.setuptools.packages.find]
include = ["settus*"]

[tool.setuptools.dynamic]
version = {attr = "settus._version.VERSION"}


[tool.ruff]
#line-length = 120
#extend-select = ['Q', 'RUF100', 'C90', 'UP', 'I', 'D', 'T']
extend-ignore = []
#flake8-quotes = {inline-quotes = 'single', multiline-quotes = 'double'}
#mccabe = { max-complexity = 14 }
#isort = { known-first-party = ['pydantic', 'tests'] }
target-version = "py39"
#extend-exclude = ['pydantic/v1', 'tests/mypy/outputs']
builtins = []

[tool.ruff.pydocstyle]
convention = "numpy"